Struct property_rs::World

pub struct World {
    // some fields omitted
}

Methods

impl World

fn new() -> World

fn add_entity(&mut self) -> Entity

Adds a new entity. Will reuse an idx if available, if not it will increment the idx counter and allocate space for the components.

fn remove_entity(&mut self, entity: Entity) -> bool

Removes an entity, but only if the uuid matches. Returns true if the entity existed and could be deleted, otherwise false.

fn is_valid_entity(&self, entity: &Entity) -> bool

Checks if an Entity reference is valid.

fn add_component<T: Any>(&mut self, entity: &Entity, component: T) -> Option<T>

Adds a new component to an entity. If the entity already had that component, that component is returned. Otherwise, None is returned.

fn get_component<T: Any>(&self, entity: &Entity) -> Option<&T>

Returns a reference to a component, if it exists.

fn get_mut_component<T: Any>(&mut self, entity: &Entity) -> Option<&mut T>

Returns a mutable reference to a component, if it exists.